added css files for components and changed file structure
