--- beer_me/src/components/BackImage.css
@@ -1,3 +1,0 @@
-.backImg {
-  height: 40px;
-}
