+++ beer_me/src/components/Layout/Layout.css
@@ -1,0 +1,20 @@
+.mid_top {
+  height: 175px;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.left_img {
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.mid_bottom {
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.right_img {
+  margin-left: auto;
+  margin-right: auto;
+}
