+++ beer_me/src/components/Searching/Searching.css
@@ -1,0 +1,20 @@
+.beer_search_img {
+  margin-left: auto;
+  margin-right: auto;
+  width: 325px;
+  height: 183px;
+}
+
+.brewery_search_img {
+  margin-left: auto;
+  margin-right: auto;
+  width: 325px;
+  height: 183px;
+}
+
+.brewery_of_month_img {
+  margin-left: auto;
+  margin-right: auto;
+  width: 325px;
+  height: 183px;
+}
